feat(models): introduce ModelsSupported component for displaying AI providers and models
